Refactoring on UUID usage in the project
